add policy

+++ app/controllers/api/v1/account/subscriptions_controller.rb
@@ -4,6 +4,8 @@ module Api
   module V1
     module Account
       class SubscriptionsController < Api::V1::Account::ApplicationController
+        after_action :verify_authorized, only: %i[destroy]
+
         def create
           subscription = Api::V1::Account::SubscriptionForm.new(params)
 
@@ -18,6 +20,7 @@ module Api
 
         def destroy
           subscription = Subscription.find(params[:id])
+          authorize([:api, :v1, :account, subscription])
 
           subscription.destroy
           render json: { status: :success }
